Working deploy on AWS lab.
Working deploy on AWS lab. This commit shows just how much has to
change for this helm deploy to work on AWS. With this the helm chart
likely will not work on my local infrastructure, but the changes here
will help illustrate what has to be refactored, broken out as config.